+++ src/main/java/isep/web/sakila/jpa/entities/Store.java
@@ -18,7 +18,7 @@ public class Store implements Serializable {
 
 	@Id
 	@Column(name="store_id", unique=true, nullable=false)
-	private int storeId;
+	private byte storeId;
 
 	@Column(name="last_update", nullable=false)
 	private Timestamp lastUpdate;
